partial of 9B2-04 | Logic Should be Moved To an Separate Function - Refactoring;
partial of STL-03 | Lack of Validation for `destination_address` on `send_payment_to_address()`
