+++ clients/vault/src/oracle/agent.rs
@@ -197,9 +197,11 @@ mod tests {
 	use serial_test::serial;
 
 	use super::*;
+	use serial_test::serial;
 
 	#[tokio::test]
 	#[ntest::timeout(1_800_000)] // timeout at 30 minutes
+	#[serial]
 	async fn test_get_proof_for_current_slot() {
 		let agent =
 			start_oracle_agent(get_test_stellar_relay_config(false), &get_test_secret_key(false))
