+++ clients/vault/src/requests/mod.rs
@@ -1,0 +1,5 @@
+pub mod execution;
+mod helper;
+mod structs;
+
+pub use structs::*;
